Added validator for daysofweek and ensured that missing params/body elements does not crash backend

+++ server/src/controllers/accountController.js
@@ -3,7 +3,9 @@ import { StatusCodes } from "http-status-codes";
 import validate from "../util/checkValidation.js";
 
 export const create = async (req, res) => {
-  validate(req);
+  if (validate(req, res)) {
+    return res;
+  }
   const { email, name, phoneNumber } = req.body;
   if (phoneNumber === null || phoneNumber === undefined) {
     await prisma.Account.create({
@@ -31,7 +33,9 @@ export const create = async (req, res) => {
 };
 
 export const login = async (req, res) => {
-  validate(req);
+  if (validate(req, res)) {
+    return res;
+  }
   const { email } = req.body;
   const account = await prisma.Account.findUnique({
     where: {
@@ -42,7 +46,9 @@ export const login = async (req, res) => {
 };
 
 export const getCourses = async (req, res) => {
-  validate(req);
+  if (validate(req, res)) {
+    return res;
+  }
   const id = parseInt(req.get("id"), 10);
   const studentCourses = await prisma.course.findMany({
     where: {
@@ -82,7 +88,9 @@ export const getCourses = async (req, res) => {
 };
 
 export const deleteAccount = async (req, res) => {
-  validate(req);
+  if (validate(req, res)) {
+    return res;
+  }
   const id = parseInt(req.get("id"), 10);
   await prisma.registration.deleteMany({
     where: {
